move style helpers to utils
