+++ app/layer-groups/boroughs.js
@@ -1,4 +1,4 @@
-import adminBoundaryStyles from '../helpers/admin-boundary-styles';
+import adminBoundaryStyles from '../utils/admin-boundary-styles';
 
 const { paint, layout } = adminBoundaryStyles;
