+++ app/layer-groups/coastal-zone-boundary.js
@@ -1,4 +1,4 @@
-import { lineStyle, fillStyle } from '../helpers/polygon-layer-styles';
+import { lineStyle, fillStyle } from '../utils/polygon-layer-styles';
 
 const legendColor = 'rgba(124, 159, 238, 1)';
 export default {
