+++ app/layer-groups/low-density-growth-mgmt-areas.js
@@ -1,4 +1,4 @@
-import { lineStyle, fillStyle } from '../helpers/polygon-layer-styles';
+import { lineStyle, fillStyle } from '../utils/polygon-layer-styles';
 
 const legendColor = 'rgba(255, 127, 0, 1)';
