+++ app/layer-groups/special-purpose-districts.js
@@ -1,4 +1,4 @@
-import { lineStyle, fillStyle } from '../helpers/polygon-layer-styles';
+import { lineStyle, fillStyle } from '../utils/polygon-layer-styles';
 
 const legendColor = 'rgba(21, 83, 30, 1)';
