+++ app/utils/polygon-layer-styles.js
