--- tests/integration/helpers/polygon-layer-styles-test.js
@@ -1,16 +1,0 @@
-
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('polygon-layer-styles', 'helper:polygon-layer-styles', {
-  integration: true
-});
-
-// Replace this with your real tests.
-test('it renders', function(assert) {
-  this.set('inputValue', '1234');
-
-  this.render(hbs`{{polygon-layer-styles inputValue}}`);
-
-  assert.equal(this.$().text().trim(), '1234');
-});
